refactor: move apply methods to model layer
