+++ packages/base-docs/src/basics/component-tools.ts
@@ -21,7 +21,7 @@ export function getDocObject(
         return;
     }
 
-    const { components, mainComponent, scene, engine } = currentRender;
+    const { mainComponent, scene, engine } = currentRender;
 
     const document = mainComponent as Documents;
