+++ packages/base-docs/src/commands/commands/inline-format.command.ts
@@ -11,9 +11,9 @@ import {
     ITextRun,
     IUndoRedoService,
     IUniverInstanceService,
+    MemoryCursor,
 } from '@univerjs/core';
 
-import { MemoryCursor } from '../../basics/memory-cursor';
 import { TextSelectionManagerService } from '../../services/text-selection-manager.service';
 import { IRichTextEditingMutationParams, RichTextEditingMutation } from '../mutations/core-editing.mutation';
