+++ packages/base-docs/src/index.ts
@@ -1,6 +1,5 @@
 export { getDocObject } from './basics/component-tools';
 export * from './basics/docs-view-key';
-export { MemoryCursor } from './basics/memory-cursor';
 export { DocCopyCommand, DocCutCommand, DocPasteCommand } from './commands/commands/clipboard.command';
 export {
     CoverCommand,
