+++ packages/base-docs/src/views/doc-canvas-view.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-magic-numbers */
 import {
     Documents,
     EVENT_TYPE,
