+++ packages/base-render/src/components/docs/block/paragraph/index.ts
@@ -1,6 +1,6 @@
 export * from './bullet';
-export * from './bullet.-ruler';
+export * from './bullet-ruler';
 export * from './inline-drawing';
-export * from './language.-ruler';
-export * from './layout.-ruler';
+export * from './language-ruler';
+export * from './layout-ruler';
 export * from './paragraph';
