+++ packages/base-render/src/components/docs/block/paragraph/language-ruler.ts
