+++ packages/core/src/common/const.ts
@@ -1,3 +1,5 @@
 export const DOCS_NORMAL_EDITOR_UNIT_ID_KEY = '__defaultDocumentNormalEditorSpecialUnitId_20231006__';
 
 export const DOCS_FORMULA_BAR_EDITOR_UNIT_ID_KEY = '__defaultDocumentFormulaBarEditorSpecialUnitId_20231012__';
+
+export const DEFAULT_EMPTY_DOCUMENT_VALUE = '\r\n';
