+++ packages/core/src/common/memory-cursor.ts
