+++ packages/core/src/docs/index.ts
@@ -1,1 +1,1 @@
-export * from './domain';
+export * from './model';
